tasks/image-region-extraction/requirements.txt
numpy
opencv-python
Pillow
coretex
